+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [5.0.0] - UNRELEASED
+### Added
+- Added support for `useRoutes` to take an array to specify route matching priority
 ### Changed
 - Added support for underscore `_` in path part matchers